Implement System.IO.Path on Unix

+++ src/System.IO.FileSystem/tests/DirectoryInfo/MoveTo_str.cs
@@ -148,24 +148,27 @@ public class DirectoryInfo_Move_str
 
             strLoc = "Loc_498vy";
 
-            dir2 = Directory.CreateDirectory(testDir);
-            iCountTestcases++;
-            try
+            if (Interop.IsWindows) // whitespace-only names are valid on Unix
             {
-                dir2.MoveTo("         ");
-                iCountErrors++;
-                printerr("Error_209uc! Expected exception not thrown");
-                fil2.Delete();
+                dir2 = Directory.CreateDirectory(testDir);
+                iCountTestcases++;
+                try
+                {
+                    dir2.MoveTo("         ");
+                    iCountErrors++;
+                    printerr("Error_209uc! Expected exception not thrown");
+                    fil2.Delete();
+                }
+                catch (ArgumentException)
+                {
+                }
+                catch (Exception exc)
+                {
+                    iCountErrors++;
+                    printerr("Error_28829! Incorrect exception thrown, exc==" + exc.ToString());
+                }
+                dir2.Delete(true);
             }
-            catch (ArgumentException)
-            {
-            }
-            catch (Exception exc)
-            {
-                iCountErrors++;
-                printerr("Error_28829! Incorrect exception thrown, exc==" + exc.ToString());
-            }
-            dir2.Delete(true);
 
             // [] Mvoe to the same directory
 
@@ -244,23 +247,26 @@ public class DirectoryInfo_Move_str
 
             strLoc = "Loc_98399";
 
-            dir2 = Directory.CreateDirectory(testDir);
-            iCountTestcases++;
-            try
+            if (Interop.IsWindows) // directory with tabs as name is valid on Unix
             {
-                dir2.MoveTo("\t");
-                iCountErrors++;
-                printerr("Error_2091c! Expected exception not thrown");
+                dir2 = Directory.CreateDirectory(testDir);
+                iCountTestcases++;
+                try
+                {
+                    dir2.MoveTo("\t");
+                    iCountErrors++;
+                    printerr("Error_2091c! Expected exception not thrown");
+                }
+                catch (ArgumentException)
+                {
+                }
+                catch (Exception exc)
+                {
+                    iCountErrors++;
+                    printerr("Error_8374v! Incorrect exception thrown, exc==" + exc.ToString());
+                }
+                dir2.Delete(true);
             }
-            catch (ArgumentException)
-            {
-            }
-            catch (Exception exc)
-            {
-                iCountErrors++;
-                printerr("Error_8374v! Incorrect exception thrown, exc==" + exc.ToString());
-            }
-            dir2.Delete(true);
